feat: Add Facebook Pixel tracking
Integrates the Meta Pixel code into the presell page to enable conversion tracking on Facebook. This allows for better audience targeting and campaign optimization.

The `AddToCart` event is now triggered for the Facebook Pixel when a user clicks the Call to Action button, in addition to the existing Kwai Pixel tracking.